Extract database encoding/decoding to a separate module
